+++ Changelog.md
@@ -1,3 +1,23 @@
+## 2.0.0-rc.24
+### Changed
+ - FIO-8106: add default storeas value to tags
+ - FIO-8106: add invalidDate error translation
+
+## 2.0.0-rc.23
+### Changed
+ - Fix: JSONLogic validations should get same context as calculations
+   
+## 2.0.0-rc.22
+### Changed
+ - FIO-7146: gh actions for repository
+ - FIO-8100: add clearhidden processor to cover logic, conditions, and custom
+ - FIO-8101: always process json validation even if value is falsy
+ - FIO-8107: correct small error in normalize processor
+   
+## 2.0.0-rc.21
+### Changed
+ - FIO-8092: update isEmpty to isComponentDataEmpty and account for differing component data types
+
 ## 2.0.0-rc.20
 ### Changed
  - FIO-8086: don't multiple validate select components